get all facilities endpoint modified, lab result report date formarted

+++ src/main/java/org/lamisplus/modules/hiv/domain/dto/TransferPatientInfo.java
@@ -18,7 +18,7 @@ public interface TransferPatientInfo {
     String getCurrentWhoClinical();
     Integer getCurrentCD4Count();
     Integer getBaselineCD4();
-    Integer getViralLoad();
+    Double getViralLoad();
     String getCurrentRegimenLine();
     String getFirstLineArtRegimen();
     Long getHivStatusId();
